Backend/bookshare/requirements.txt
altgraph==0.16.1
apturl==0.5.2
asgiref==3.2.7
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
Brlapi==0.6.6
certifi==2019.3.9
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.10
decorator==4.4.1
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
entrypoints==0.3
future==0.17.1
gym==0.12.1
httplib2==0.9.2
idna==2.8
importlib-metadata==1.3.0
IPy==0.83
ipykernel==5.1.3
ipython==7.11.0
ipython-genutils==0.2.0
isort==4.3.20
jedi==0.15.2
Jinja2==2.10.3
joblib==0.13.2
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
Lasagne==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.1
louis==3.5.0
macaroonbakery==1.1.3
macholib==1.11
Mako==1.0.7
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==4.4.0
netifaces==0.10.4
networkx==1.11
notebook==6.0.2
numpy==1.18.0
oauth==1.0.1
olefile==0.45.1
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pefile==2019.4.18
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.0.0
psycopg2==2.8.5
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygame==1.9.6
pyglet==1.3.2
Pygments==2.5.2
pygobject==3.26.1
PyInstaller==3.4
pylint==2.3.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.0
pyRFC3339==1.0
pyrsistent==0.15.6
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
pyzmq==18.1.1
reportlab==3.4.0
requests==2.21.0
requests-unixsocket==0.1.5
scikit-learn==0.21.2
scikit-neuralnetwork==0.7
scipy==1.3.0
scour==0.36
SecretStorage==2.3.1
Send2Trash==1.5.0
sepolicy==1.1
setools==4.1.1
simplejson==3.13.2
six==1.13.0
sklearn==0.0
sqlparse==0.3.1
system-service==0.3
systemd-python==234
terminado==0.8.3
testpath==0.4.4
Theano==1.0.4
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.2
usb-creator==0.3.3
vboxapi==1.0
virtualenv==15.1.0
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.11.1
xkit==0.0.0
zipp==0.6.0
zope.interface==4.3.2
